relocate logo to assets folder
