+++ frontend/src/components/Signup.tsx
@@ -1,4 +1,4 @@
-import tiger from './tiger.png';
+import tiger from '../assets/logo.png';
 
 function Signup() {
   return (
